Added Credentials and updated User classes

+++ src/main/java/com/cooksystems/springassessmentsocialmediaapr2024team3/dtos/UserResponseDto.java
@@ -1,6 +1,6 @@
 package com.cooksystems.springassessmentsocialmediaapr2024team3.dtos;
 
-import java.security.Timestamp;
+import java.sql.Timestamp;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,21 +8,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class UserResponseDto {
-    private Long id;
-
+    private Integer id;
     private String username;
-
     private String password;
-
     private Timestamp joined; 
-
     private boolean deleted;
-
     private String firstName;
-
     private String lastName;
-
     private String email;
-
     private String phone;
 }
